refactor(main): decompose web-server-factory.ts into per-domain callback modules
refactor(main): decompose web-server-factory.ts into per-domain callback modules

Split the 3,432-line web-server-factory.ts into a thin orchestrator plus
23 domain files under src/main/web-server/callbacks/, mirroring the
messageHandlers/routes/managers/services split already used elsewhere
in the web-server layer.

- No behavior change, verified via a completeness diff of every
  server.setXCallback(...) registration against the pre-refactor file
- Added smoke tests for the domains that previously had zero coverage
- Fixed __COMMIT_HASH__ not being defined in vite.config.web-desktop.mts
- Updated docs/agent-guides/PROCESS-SYSTEM.md and WEB-MOBILE.md for the
  new callbacks/ directory structure
