+++ docs/agent-guides/WEB-MOBILE.md
@@ -61,7 +61,7 @@ The web-desktop build aliases `electron` to `src/web-desktop/electron-shim.ts` i
 
 - `contextBridge.exposeInMainWorld('maestro', ...)` writes to `window.maestro` in the browser.
 - `ipcRenderer.invoke(channel, ...args)` becomes a `bridge.invoke` WebSocket frame to `/$TOKEN/ws`, resolved by the main process and returned over the same socket.
-- Main -> renderer push events reach browser clients through `safeSend` (`src/main/utils/safe-send.ts`), which fans each event out to the desktop `webContents` AND to the bridge via `broadcastBridgeEvent`. (One deliberate exception is documented in [Deferred: web-server-factory.ts](#deferred-web-server-factoryts).)
+- Main -> renderer push events reach browser clients through `safeSend` (`src/main/utils/safe-send.ts`), which fans each event out to the desktop `webContents` AND to the bridge via `broadcastBridgeEvent`. (One deliberate exception is documented in [Deferred: web-server/callbacks/\*.ts](#deferred-web-servercallbacksts).)
 
 This is why the renderer's own Zustand stores, IPC service wrappers, and components all work in the browser with no web-specific fork.
 
@@ -159,13 +159,13 @@ The web-desktop Vite config sets `publicDir: src/web/public`, so a `build:web-de
 
 ---
 
-## Deferred: web-server-factory.ts
+## Deferred: web-server/callbacks/\*.ts
 
 Main-to-renderer push events only reach browser clients when they go through `safeSend` (`src/main/utils/safe-send.ts`), which fans each event out to the desktop `webContents` AND to the web-desktop bridge via `broadcastBridgeEvent`. The Phase 01 migration routed the session/app data sends across the IPC handlers through `safeSend` so web clients stop silently missing group chat, stats, Cue, and Auto Run events.
 
-`src/main/web-server/web-server-factory.ts` was deliberately left out of that migration. Its ~58 direct `webContents.send(...)` calls are not new events originating in the main process: they mirror actions that a web client already performed (over the WebSocket bridge) back onto the desktop renderer so the two surfaces stay in sync. Bridging those sends through `safeSend` would echo each web-originated action straight back to the web client that initiated it, causing duplicate state updates and feedback loops.
+`src/main/web-server/web-server-factory.ts` and the per-domain callback modules it wires up (`src/main/web-server/callbacks/*.ts`) were deliberately left out of that migration. Their ~60 direct `webContents.send(...)` calls are not new events originating in the main process: they mirror actions that a web client already performed (over the WebSocket bridge) back onto the desktop renderer so the two surfaces stay in sync. Bridging those sends through `safeSend` would echo each web-originated action straight back to the web client that initiated it, causing duplicate state updates and feedback loops.
 
-Wiring the factory into the bridge therefore requires an echo-suppression design (for example, tagging each mirrored event with its originating client id and having the bridge skip re-delivering it to that origin) before the sends can safely fan out. That work is out of scope for the safeSend parity pass and is tracked as a separate effort. Until then, leave the `web-server-factory.ts` sends as direct `webContents.send(...)` calls.
+Wiring the factory into the bridge therefore requires an echo-suppression design (for example, tagging each mirrored event with its originating client id and having the bridge skip re-delivering it to that origin) before the sends can safely fan out. That work is out of scope for the safeSend parity pass and is tracked as a separate effort. Until then, leave the `web-server-factory.ts` / `callbacks/*.ts` sends as direct `webContents.send(...)` calls.
 
 ---
